Adds blockchain synchronization tracker
This will allow for faster blockchain synchronization times, as the
whole blockchain won't have to be checked every time.

Next step would be to add soft fork resilience. This can also be
achieved by the software that uses cloudmanager, if it only marks a
block as visited after N confirmations.